Add files via upload
This short notebook produces a word cloud in Python from a text file of several job descriptions that has the capability to show phrases depending on the collocation threshold set. In other words, if the collocation threshold is set higher, then the word cloud will mostly display single words. If the threshold is set to numbers less than 10, the word cloud will display mostly phrases. Additionally, a complete list of important word combinations will be provided.